refactor: 사진이 없는 게시글 삭제 수정

+++ server/src/main/java/mainproject/cafeIn/domain/post/service/PostService.java
@@ -117,7 +117,10 @@ public class PostService {
 
         Long cafeId = findPost.getCafe().getId();
 
-        imageService.delete("posts", findPost.getImage());
+
+        if (findPost.getImage() != null) {
+            imageService.delete("posts", findPost.getImage());
+        }
 
         postRepository.delete(findPost);
